Add ActionsDropdownDecorator to Text; remove TextFieldWithActions
