+++ app/src/shared/components/TextField/textField.pcss
@@ -106,15 +106,3 @@
 .numberButton:first-child {
   border-bottom: 1px solid #EFEFEF !important;
 }
-
-.actionContainer {
-  display: inline-block;
-  position: absolute;
-  right: 0.5rem;
-  top: 50%;
-  transform: translate(0, -50%);
-}
-
-.actionContainer.isOpen {
-  z-index: 2;
-}
